error in gst tax.Q ..need to fix this bug

+++ PROGRAMS/gst_tax.java
@@ -11,7 +11,8 @@ public class gst_tax{
         float pen=sc.nextFloat();   
         System.out.println("Enter the price for the erasor : ");
         float erasor=sc.nextFloat();
-        float total=(pencil+pen+erasor+ (0.18*pencil)+(0.18*pen)+(0.18*erasor));
+        float total=float(pencil+pen+erasor+ (0.18*pencil)+(0.18*pen)+(0.18*erasor));
         System.out.println("Total cost of the products are : "+total);
+        
     }
 }
